Fix "Energy" values + small updates
Huge fix regarding "Energy/Forward" and "Energy/Reverse".
Values needs to be a total, so that the vrmlogger in VenusOS can get a delta with a change and update VRM statistics correct.

Also changed some formatting + ProductId of the Shelly - we are emulating a ET340 now

Special thanks to @vikt0rm how also pointed out the Energy fix, almost at the time I found it by myself :+1:

Co-Authored-By: vikt0rm <39840167+vikt0rm@users.noreply.github.com>